+++ src/features/recurring-bills/data-access.ts
@@ -2,7 +2,6 @@ import "server-only"
 
 import { redirect } from "next/navigation"
 
-import { verifySession } from "@/data-access/auth"
 import {
   getBillMonthlyStatus,
   getDaysUntilDue,
@@ -12,8 +11,14 @@ import {
   type BillMonthlyStatus,
 } from "@/features/recurring-bills/helpers"
 import { RecurringBillCreate } from "@/features/recurring-bills/schemas"
+import { DEMO_ACCOUNT_ERROR_MESSAGE } from "@/lib/constants"
 import { prisma, type Prisma } from "@/lib/prisma"
-import { type RecurringBillCreateErrors, type DALReturn } from "@/lib/types"
+import { verifySession } from "@/lib/session"
+import {
+  type DALReturn,
+  type DALDeleteItemReurn,
+  type RecurringBillCreateErrors,
+} from "@/lib/types"
 
 // ============================================
 // =========== Create Recurring Bill ==========
@@ -25,13 +30,13 @@ export async function createRecurringBill({
 }: RecurringBillCreate): Promise<
   DALReturn<RecurringBillCreateErrors> & { recurringBillId?: string }
 > {
-  const userId = await verifySession()
-  if (!userId) redirect("/login")
+  const session = await verifySession()
+  if (!session) redirect("/login")
 
   try {
     const result = await prisma.recurringBill.create({
       data: {
-        userId,
+        userId: session.userId,
         counterparty,
         amount,
       },
@@ -88,8 +93,8 @@ export async function getRecurringBills({
   query = "",
   sortby = "daysAsc",
 }: GetRecurringBillsParams) {
-  const userId = await verifySession()
-  if (!userId) redirect("/login")
+  const session = await verifySession()
+  if (!session) redirect("/login")
 
   // await new Promise((resolve) => setTimeout(resolve, 3000))
 
@@ -122,13 +127,13 @@ export async function getRecurringBills({
   const [allRecurringBills, filteredRecurringBills, recurringBillsSummary] =
     await prisma.$transaction([
       prisma.recurringBill.findMany({
-        where: { userId },
+        where: { userId: session.userId },
         select: recurringBillSelect,
       }),
 
       prisma.recurringBill.findMany({
         where: {
-          userId,
+          userId: session.userId,
           counterparty: { contains: query, mode: "insensitive" },
         },
         orderBy: orderBy,
@@ -136,7 +141,7 @@ export async function getRecurringBills({
       }),
 
       prisma.recurringBill.aggregate({
-        where: { userId },
+        where: { userId: session.userId },
         _sum: { amount: true },
         _count: { _all: true },
       }),
@@ -197,15 +202,19 @@ export async function getRecurringBills({
 
 export async function deleteRecurringBill(
   id: string
-): Promise<{ success: boolean }> {
-  const userId = await verifySession()
-  if (!userId) redirect("/login")
+): Promise<DALDeleteItemReurn> {
+  const session = await verifySession()
+  if (!session) redirect("/login")
+
+  if (session.role === "DEMO") {
+    return { success: false, message: DEMO_ACCOUNT_ERROR_MESSAGE }
+  }
 
   try {
-    await prisma.recurringBill.delete({ where: { id, userId } })
+    await prisma.recurringBill.delete({ where: { id, userId: session.userId } })
     return { success: true }
   } catch (e) {
     console.error(e)
-    return { success: false }
+    return { success: false, message: "Error deleting bill. Please try agian." }
   }
 }
